Add event listeners for operands, fix storeVal()
Need to be able to store value of first number in operation and the
operand. Fix storeVal function so that it stores these values and gets
ready to accept a new number. Add event listeners for each operand
button that store this information in the equation object if there isn't
already an operand value.

Still need to introduce the calculate function.